commands: move log code to log.rs
